+++ src/rss/api/app.py
@@ -1,11 +1,9 @@
-import os.path
-
-from flask import current_app, request, jsonify, Response
+from flask import Response
 from flask_httpauth import HTTPBasicAuth
 
 from rss.api import API_CONFIG, create_app, db
-from rss.api import errors
-from rss.api.models import Alert, query_alerts
+from rss.api.errors import errors
+from rss.api.alerts.models import Alert
 from rss.utils.wait_for_db import wait_for_postgres
 
 
@@ -20,7 +18,7 @@ def verify_password(user: str, password: str) -> bool:
 
 
 @app.errorhandler(400)
-def bad_request(error):
+def bad_request(error) -> Response:
     return errors.bad_request("Invalid request")
 
 
@@ -30,17 +28,17 @@ def unauthorized(error) -> Response:
 
 
 @app.errorhandler(403)
-def forbidden(error):
+def forbidden(error) -> Response:
     return errors.forbidden("Forbidden")
 
 
 @app.errorhandler(404)
-def page_not_found(error):
+def page_not_found(error) -> Response:
     return errors.not_found("The requested resource could not be found")
 
 
 @app.errorhandler(500)
-def internal_server_error(error):
+def internal_server_error(error) -> Response:
     return errors.bad_request("There was an error handling your request")
 
 
@@ -49,85 +47,9 @@ def make_shell_context():
     return dict(db=db, Alert=Alert)
 
 
-@app.route("/")
+@app.route("/ping")
 def index():
-    return "OK", 200
-
-
-@app.route(f"{api_route}/alerts/", methods=["POST"])
-@auth.login_required
-def add_new_alert():
-    id_ = request.json["id"]
-    if Alert.get_by_identifier(id_) is not None:
-        return errors.bad_request(f"Alert with identifier {id_} already exists")
-
-    alert = Alert.from_json(request.json)
-    save_path = request.args.get("save_path", "")
-
-    if save_path and os.path.isdir(save_path):
-        alert.save_to_file(save_path, current_app.logger)
-    db.session.commit()
-    return jsonify(alert.to_json()), 201
-
-
-@app.route(f"{api_route}/alerts/<identifier>")
-def get_alert(identifier):
-    alert = Alert.get_by_identifier(identifier)
-    if alert is not None:
-        return jsonify(alert.to_json())
-    return errors.not_found(f"Alert with identifier {identifier} could not be found")
-
-
-@app.route(f"{api_route}/alerts/")
-def get_alerts():
-    alert_type = request.args.get("type", "all")
-    page = request.args.get("page", 1, type=int)
-    start_date = request.args.get("start_date", "")
-    end_date = request.args.get("end_date", "")
-    region = request.args.get("region", "")
-    state = request.args.get("state", "")
-
-    if alert_type not in {"all", "alert", "event"}:
-        return errors.bad_request(f"Invalid alert type {alert_type}")
-
-    alerts, prev, next_page, total = query_alerts(
-        page, alert_type, start_date, end_date, region, state)
-    return jsonify({
-        "alerts": [al.to_json() for al in alerts],
-        "prev": prev,
-        "next": next_page,
-        "count": total,
-    })
-
-
-@app.route(f"{api_route}/alerts/<identifier>/cap/")
-def get_cap_file(identifier):
-    save_file = request.args.get("save", "false")
-
-    alert = Alert.get_by_identifier(identifier)
-    if alert is not None:
-        file_contents = alert.to_cap_file()
-        if save_file.lower() == "true":
-            if identifier == "latest":
-                identifier = "sasmex"
-            response = Response(file_contents, mimetype="text/xml")
-            response.headers.set(
-                "Content-Disposition", "attachment", filename=f"{identifier}.cap"
-            )
-            return response
-        else:
-            return jsonify({
-                "contents": file_contents
-            })
-    return errors.not_found(f"Alert with identifier {identifier} could not be found")
-
-
-@app.route(f"{api_route}/alerts/latest/")
-def get_last_alert():
-    alert = Alert.get_by_identifier("latest")
-    if alert is not None:
-        return jsonify(alert.to_json())
-    return errors.not_found("No latest alerts found")
+    return {"message": "ok"}
 
 
 @app.cli.command("wait-for-db")